V2 options applying refactoring (#1881)
* presenter skeleton

* test fix

* rm presentation

* options presenter

* test refactoring
