+++ lib/android/app/src/main/java/com/reactnativenavigation/viewcontrollers/ContainerViewController.java
@@ -5,6 +5,7 @@ import android.support.annotation.NonNull;
 import android.view.View;
 
 import com.reactnativenavigation.parse.NavigationOptions;
+import com.reactnativenavigation.presentation.OptionsPresenter;
 
 public class ContainerViewController extends ViewController {
 
@@ -28,16 +29,17 @@ public class ContainerViewController extends ViewController {
 	}
 
 	private final String containerName;
+	private OptionsPresenter optionsPresenter;
 
 	private final ContainerViewCreator viewCreator;
 	private final NavigationOptions navigationOptions;
 	private ContainerView containerView;
 
 	public ContainerViewController(final Activity activity,
-	                               final String id,
-	                               final String containerName,
-	                               final ContainerViewCreator viewCreator,
-	                               final NavigationOptions initialNavigationOptions) {
+								   final String id,
+								   final String containerName,
+								   final ContainerViewCreator viewCreator,
+								   final NavigationOptions initialNavigationOptions) {
 		super(activity, id);
 		this.containerName = containerName;
 		this.viewCreator = viewCreator;
@@ -55,7 +57,7 @@ public class ContainerViewController extends ViewController {
 	public void onViewAppeared() {
 		super.onViewAppeared();
 		ensureViewIsCreated();
-		applyNavigationOptions();
+		applyOptions();
 		containerView.sendContainerStart();
 	}
 
@@ -79,19 +81,15 @@ public class ContainerViewController extends ViewController {
 
 	public void mergeNavigationOptions(final NavigationOptions options) {
 		navigationOptions.mergeWith(options);
-		applyNavigationOptions();
-	}
-
-	private void applyNavigationOptions() {
-		if (getParentStackController() != null) {
-			getParentStackController().getTopBar().setTitle(navigationOptions.title);
-			getParentStackController().getTopBar().setBackgroundColor(navigationOptions.topBarBackgroundColor);
-			getParentStackController().getTopBar().setTitleTextColor(navigationOptions.topBarTextColor);
-			getParentStackController().getTopBar().setTitleFontSize(navigationOptions.topBarTextFontSize);
-		}
+		applyOptions();
 	}
 
 	public NavigationOptions getNavigationOptions() {
 		return navigationOptions;
+	}
+
+	private void applyOptions() {
+		OptionsPresenter presenter = new OptionsPresenter(getParentStackController());
+		presenter.applyOptions(navigationOptions);
 	}
 }
